Fix makepri
Thanks to https://github.com/MustardChef/WSABuilds

+++ installer/x64/MakePri.ps1
@@ -16,20 +16,25 @@
 # Copyright (C) 2023 LSPosed Contributors
 #
 
-$MakePri = "C:\Program Files (x86)\Windows Kits\10\bin\10.0.22621.0\x64\makepri.exe"
+$MakePri = Join-Path $PSScriptRoot "makepri.exe"
 
 New-Item -Path "." -Name "priinfo" -ItemType "directory" | Out-Null
-Copy-Item .\resources.pri -Destination ".\pri\resources.pri" | Out-Null
-$AppxManifestFile = ".\AppxManifest.xml"
+if (Test-Path .\resources.pri) {
+    New-Item -Path "." -Name "pri" -ItemType "directory" -Force | Out-Null
+    Copy-Item .\resources.pri -Destination ".\pri\resources.pri" | Out-Null
+}
+$AppxManifestFile = Join-Path $PSScriptRoot "AppxManifest.xml"
 $PriItem = Get-Item ".\pri\*" -Include "*.pri"
 Write-Output "Dumping resources..."
-foreach ($Item in $PriItem) {
-    Start-Process -PassThru $MakePri -Args "dump /if $($Item | Resolve-Path -Relative) /o /es .\pri\resources.pri /of .\priinfo\$($Item.Name).xml /dt detailed"
+$Processes = foreach ($Item in $PriItem) {
+    Start-Process -PassThru -WindowStyle Hidden $MakePri -Args "dump /if $($Item | Resolve-Path -Relative) /o /es .\pri\resources.pri /of .\priinfo\$($Item.Name).xml /dt detailed"
+}
+if ($Processes) {
+    $Processes | Wait-Process
 }
 
 Write-Output "Creating pri from dumps...."
-$ProcNewFromDump = Start-Process -PassThru $MakePri -Args "new /pr .\priinfo /cf .\xml\priconfig.xml /of .\resources.pri /mn $AppxManifestFile /o"
-$null = $ProcNewFromDump.Handle
+$ProcNewFromDump = Start-Process -PassThru -NoNewWindow $MakePri -Args "new /pr .\priinfo /cf .\xml\priconfig.xml /of .\resources.pri /mn $AppxManifestFile /o"
 $ProcNewFromDump.WaitForExit()
 Remove-Item 'priinfo' -Recurse -Force
 if ($ProcNewFromDump.ExitCode -ne 0) {
@@ -48,6 +53,7 @@ $ProjectXml.Save($AppxManifestFile)
 
 Remove-Item 'pri' -Recurse -Force
 Remove-Item 'xml' -Recurse -Force
-Remove-Item 'filelist-pri.txt' -Force
+Remove-Item (Join-Path $PSScriptRoot 'makepri.exe') -Force
+Remove-Item (Join-Path $PSScriptRoot 'filelist-pri.txt') -Force
 Remove-Item $PSCommandPath -Force
 exit 0
